refactor get started and auth code
