+++ lib/core/utils/app_assets.dart
@@ -1,0 +1,25 @@
+// ignore_for_file: prefer_single_quotes
+class Assets {
+  Assets._();
+  
+  /// Assets for images1
+  /// assets/images/1.png
+  static const String images1 = "assets/images/1.png";
+
+  /// Assets for imagesBags
+  /// assets/images/bags.svg
+  static const String imagesBags = "assets/images/bags.svg";
+
+  /// Assets for imagesBanner
+  /// assets/images/banner.png
+  static const String imagesBanner = "assets/images/banner.png";
+
+  /// Assets for imagesFacebook
+  /// assets/images/facebook.svg
+  static const String imagesFacebook = "assets/images/facebook.svg";
+
+  /// Assets for imagesGoogle
+  /// assets/images/google.svg
+  static const String imagesGoogle = "assets/images/google.svg";
+}
+
