+++ lib/core/utils/app_colors.dart
@@ -1,0 +1,13 @@
+import 'dart:ui';
+
+abstract class AppColors {
+  static const Color primary = Color(0xFFEF3651);
+  static const Color gray = Color(0xFFABB4BD);
+  static const Color background = Color(0xFF1E1F28);
+  static const Color dark = Color(0xFF2A2C36);
+  static const Color error = Color(0xFFFF2424);
+  static const Color success = Color(0xFF55D85A);
+  static const Color saleOrHot = Color(0xFFFF3E3E);
+  static const Color ordinaryText = Color(0xFFF5F5F5);
+  static const Color whiteColor = Color(0xFFFFFFFF);
+}
