+++ lib/features/auth/presentation/manager/check_email_forget_password_bloc/check_email_forget_password_event.dart
@@ -1,0 +1,13 @@
+part of 'check_email_forget_password_bloc.dart';
+
+
+
+// Events
+abstract class CheckEmailForgetPasswordEvent {}
+
+class EmailChanged extends CheckEmailForgetPasswordEvent {
+  final String email;
+  EmailChanged(this.email);
+}
+
+class ForgotPasswordSubmitted extends CheckEmailForgetPasswordEvent {}
