+++ lib/features/auth/presentation/manager/login_bloc/login_event.dart
@@ -1,0 +1,30 @@
+part of 'login_bloc.dart';
+
+abstract class LoginEvent extends Equatable {
+  const LoginEvent();
+
+  @override
+  List<Object?> get props => [];
+}
+
+class EmailChanged extends LoginEvent {
+  final String email;
+
+  const EmailChanged(this.email);
+
+  @override
+  List<Object?> get props => [email];
+}
+
+class PasswordChanged extends LoginEvent {
+  final String password;
+
+  const PasswordChanged(this.password);
+
+  @override
+  List<Object?> get props => [password];
+}
+
+class ShowPassword extends LoginEvent {}
+
+class LoginSubmitted extends LoginEvent {}
