+++ lib/features/auth/presentation/manager/login_bloc/login_state.dart
@@ -1,0 +1,76 @@
+part of 'login_bloc.dart';
+
+class LoginState extends Equatable {
+  final String email;
+  final String password;
+  final bool isEmailValid;
+  final bool isPasswordValid;
+  final bool isSubmitting;
+  final bool isSuccess;
+  final bool isFailure;
+  final bool showOrHidePass;
+
+  const LoginState({
+    required this.email,
+    required this.showOrHidePass,
+    required this.password,
+    required this.isEmailValid,
+    required this.isPasswordValid,
+    required this.isSubmitting,
+    required this.isSuccess,
+    required this.isFailure,
+  });
+
+  bool get isFormValid => isEmailValid && isPasswordValid;
+
+  factory LoginState.initial() {
+    return const LoginState(
+      email: '',
+      password: '',
+      isEmailValid: true,
+      isPasswordValid: true,
+      isSubmitting: false,
+      showOrHidePass: false,
+      isSuccess: false,
+      isFailure: false,
+    );
+  }
+
+  LoginState copyWith({
+    String? email,
+    String? password,
+    bool? isEmailValid,
+    bool? isPasswordValid,
+    bool? isSubmitting,
+    bool? isSuccess,
+    bool? showOrHidePass,
+    bool? isFailure,
+  }) {
+    return LoginState(
+      email: email ?? this.email,
+      showOrHidePass: showOrHidePass ?? this.showOrHidePass,
+      password: password ?? this.password,
+      isEmailValid: isEmailValid ?? this.isEmailValid,
+      isPasswordValid: isPasswordValid ?? this.isPasswordValid,
+      isSubmitting: isSubmitting ?? this.isSubmitting,
+      isSuccess: isSuccess ?? this.isSuccess,
+      isFailure: isFailure ?? this.isFailure,
+    );
+  }
+
+  @override
+  List<Object?> get props => [
+        email,
+        password,
+        showOrHidePass,
+        isEmailValid,
+        isPasswordValid,
+        isSubmitting,
+        isSuccess,
+        isFailure,
+      ];
+
+      
+}
+
+
