+++ lib/features/auth/presentation/manager/sign_up_bloc/sign_up_event.dart
@@ -1,0 +1,22 @@
+part of 'sign_up_bloc.dart';
+
+
+// Events
+abstract class SignUpEvent {}
+
+class NameChanged extends SignUpEvent {
+  final String name;
+  NameChanged(this.name);
+}
+
+class EmailChanged extends SignUpEvent {
+  final String email;
+  EmailChanged(this.email);
+}
+
+class PasswordChanged extends SignUpEvent {
+  final String password;
+  PasswordChanged(this.password);
+}
+
+class SignUpSubmitted extends SignUpEvent {}
