+++ lib/features/auth/presentation/manager/sign_up_bloc/sign_up_state.dart
@@ -1,0 +1,64 @@
+part of 'sign_up_bloc.dart';
+
+// States
+class SignUpState {
+  final String name;
+  final String email;
+  final String password;
+  final bool isNameValid;
+  final bool isEmailValid;
+  final bool isPasswordValid;
+  final bool isSubmitting;
+  final bool isSuccess;
+  final bool isFailure;
+
+  SignUpState({
+    required this.name,
+    required this.email,
+    required this.password,
+    required this.isNameValid,
+    required this.isEmailValid,
+    required this.isPasswordValid,
+    required this.isSubmitting,
+    required this.isSuccess,
+    required this.isFailure,
+  });
+
+  factory SignUpState.initial() {
+    return SignUpState(
+      name: '',
+      email: '',
+      password: '',
+      isNameValid: true,
+      isEmailValid: true,
+      isPasswordValid: true,
+      isSubmitting: false,
+      isSuccess: false,
+      isFailure: false,
+    );
+  }
+
+  SignUpState copyWith({
+    String? name,
+    String? email,
+    String? password,
+    bool? isNameValid,
+    bool? isEmailValid,
+    bool? isPasswordValid,
+    bool? isSubmitting,
+    bool? isSuccess,
+    bool? isFailure,
+  }) {
+    return SignUpState(
+      name: name ?? this.name,
+      email: email ?? this.email,
+      password: password ?? this.password,
+      isNameValid: isNameValid ?? this.isNameValid,
+      isEmailValid: isEmailValid ?? this.isEmailValid,
+      isPasswordValid: isPasswordValid ?? this.isPasswordValid,
+      isSubmitting: isSubmitting ?? this.isSubmitting,
+      isSuccess: isSuccess ?? this.isSuccess,
+      isFailure: isFailure ?? this.isFailure,
+    );
+  }
+}
