+++ lib/features/auth/presentation/view/widget/forget_password_button.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_ecommerce_app_ui/core/utils/app_colors.dart';
+import 'package:flutter_ecommerce_app_ui/core/utils/app_styles.dart';
+
+class ForgetPasswordButton extends StatelessWidget {
+  const ForgetPasswordButton({
+    super.key,
+    this.onPressed,
+    required this.text,
+  });
+  final Function()? onPressed;
+  final String text;
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.end,
+      children: [
+        TextButton(
+          onPressed: onPressed,
+          child: Text(
+            text,
+            style: AppStyle.styleBold14(context),
+          ),
+        ),
+        const Icon(
+          Icons.arrow_right_alt,
+          color: AppColors.primary,
+        )
+      ],
+    );
+  }
+}
