+++ lib/features/get_started/presentation/view/page/get_started_view.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_ecommerce_app_ui/features/get_started/presentation/view/widget/get_started_view_body.dart';
+
+class GetStartedView extends StatelessWidget {
+  const GetStartedView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: GetStartedViewBody(),
+    );
+  }
+}
