--- lib/ui/forgotPassword/forgot_password_screen.dart
@@ -1,98 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_ecommerce_app_ui/core/routes/app_routes.dart';
-import 'package:flutter_ecommerce_app_ui/utils/colors.dart';
-import 'package:go_router/go_router.dart';
-
-import 'forgot_password_bloc.dart';
-
-class ForgotPasswordScreen extends StatelessWidget {
-  const ForgotPasswordScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocProvider(
-      create: (context) => ForgotPasswordBloc(),
-      child: Scaffold(
-        backgroundColor: background,
-        resizeToAvoidBottomInset: true,
-        appBar: AppBar(backgroundColor: background),
-        body: SingleChildScrollView(
-          child: Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: BlocBuilder<ForgotPasswordBloc, ForgotPasswordState>(
-              builder: (context, state) {
-                return Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    const Text(
-                      "Forgot password",
-                      style: TextStyle(
-                        fontSize: 28,
-                        fontWeight: FontWeight.bold,
-                        color: Colors.white,
-                      ),
-                    ),
-                    const SizedBox(height: 20),
-                    const Text(
-                      "Please, enter your email address. You will receive a link to create a new password via email.",
-                      style: TextStyle(
-                        color: Colors.white,
-                      ),
-                    ),
-                    const SizedBox(height: 40),
-                    TextField(
-                      onChanged: (value) => context
-                          .read<ForgotPasswordBloc>()
-                          .add(EmailChanged(value)),
-                      decoration: InputDecoration(
-                        filled: true,
-                        fillColor: dark,
-                        labelText: 'Email',
-                        labelStyle: const TextStyle(color: Colors.white),
-                        suffixIcon: !state.isEmailValid
-                            ? const Icon(Icons.close, color: Colors.red)
-                            : null,
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(10),
-                        ),
-                        errorText: !state.isEmailValid
-                            ? "Not a valid email address. Should be your@email.com"
-                            : null,
-                      ),
-                      style: const TextStyle(color: Colors.white),
-                    ),
-                    const SizedBox(height: 40),
-                    state.isSubmitting
-                        ? const Center(child: CircularProgressIndicator())
-                        : ElevatedButton(
-                            onPressed: () {
-                              context.go(AppRoutes.getStarted);
-                              // context
-                              //     .read<ForgotPasswordBloc>()
-                              //     .add(ForgotPasswordSubmitted());
-                            },
-                            style: ElevatedButton.styleFrom(
-                              backgroundColor: primary,
-                              padding: const EdgeInsets.symmetric(vertical: 15),
-                              shape: RoundedRectangleBorder(
-                                borderRadius: BorderRadius.circular(30),
-                              ),
-                              minimumSize: const Size(double.infinity, 50),
-                            ),
-                            child: const Text(
-                              "SEND",
-                              style:
-                                  TextStyle(fontSize: 16, color: Colors.white),
-                            ),
-                          ),
-                  ],
-                );
-              },
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
