--- lib/utils/colors.dart
@@ -1,10 +1,0 @@
-import 'dart:ui';
-
-const Color primary = Color(0xFFEF3651);
-const Color gray = Color(0xFFABB4BD);
-const Color background = Color(0xFF1E1F28);
-const Color dark = Color(0xFF2A2C36);
-const Color error = Color(0xFFFF2424);
-const Color success = Color(0xFF55D85A);
-const Color saleOrHot = Color(0xFFFF3E3E);
-const Color ordinaryText = Color(0xFFF5F5F5);
