--- lib/utils/styles.dart
@@ -1,29 +1,0 @@
-import 'package:flutter/cupertino.dart';
-
-abstract class Styles {
-  static const textStyle18 = TextStyle(
-    fontSize: 18,
-    fontWeight: FontWeight.w600,
-  );
-  static const textStyle20 = TextStyle(
-    fontSize: 20,
-    fontWeight: FontWeight.normal,
-  );
-  static const textStyle30 = TextStyle(
-    fontSize: 30,
-    fontWeight: FontWeight.w900,
-    //fontFamily: kGtSectraFine,
-    letterSpacing: 1.2,
-  );
-  static const textStyle14 = TextStyle(
-    fontSize: 14,
-    fontWeight: FontWeight.normal,
-  );
-
-  static const textStyle16 = TextStyle(
-    fontSize: 16,
-    fontWeight: FontWeight.w500,
-  );
-}
-
-const TextStyle textStyle = TextStyle();
